+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ### Added:
-- 
+- Added defaultCommonRareness option
+- Added displayEmptyLineOnNoneRareness option
 ### Fixed:
 - 
 ### Changed:
-- Updated to mc 1.20.1